User, Customer düzenlendi, Rental'da hesaplama eklendi

+++ rentAcar/src/main/java/com/turkcell/rentAcar/business/requests/additionalService/CreateAdditionalServiceRequest.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CreateAdditionalServiceRequest {
 	private String name;
-	private int dailyPrice;
+	private double dailyPrice;
 }
